change: moved ConnectDB function to utils module
