+++ handlers/stats.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/mergestat/timediff"
 	"github.com/wavly/shawty/asserts"
-	"github.com/wavly/shawty/database"
 	sqlc "github.com/wavly/shawty/sqlc_db"
 	"github.com/wavly/shawty/utils"
 )
@@ -29,7 +28,7 @@ func Stats(w http.ResponseWriter, r *http.Request) {
 	}
 
 	templ := template.Must(template.ParseFiles("./templs/stat.html"))
-	db := database.ConnectDB()
+	db := utils.ConnectDB()
 	defer db.Close()
 	queries := sqlc.New(db)
